Move flatmap knowledge functions to new file
